+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Faulty error reporting in Shopify container
 
+### Changed
+
+- Extend timeout for Shopify webhook to 60s
+
 ## [1.7.1] - 2025-10-06
 
 ### Changed